feat: Lógica para entrada de novos usuários

+++ api/src/services/TransacaoServices/DeleteTransacaoService.ts
@@ -5,10 +5,10 @@ import Transacoes from "../../models/Transacoes";
 
 
 
-const DeleteTransacaoService = async (id: string) => {
-    const transacao = await Transacoes.findByPk(id);
+const DeleteTransacaoService = async (id: string, userId: string) => {
+    const transacao = await Transacoes.findOne({ where: { id, userId } });
     if (!transacao) throw new AppError('Transação não encontrado!');
-    Transacoes.destroy({ where: { id } })
+    Transacoes.destroy({ where: { id, userId } });
 }
 
 export default DeleteTransacaoService;
